Separate logic and DOm functions
