+++ src/index.js
@@ -1,3 +1,4 @@
+import {deleteEl, processData, errorhandle} from './logic'
 import './style.css';
 import Rainy from './img/rainy.jpg';
 import Snowy from './img/snowy.jpg';
@@ -12,11 +13,6 @@ const image = new Image();
 image.src = Rainy;
 body.style.backgroundImage = `url(${Rainy})`;
 
-function deleteEl(element) {
-  while (element.firstChild) {
-    element.removeChild(element.firstChild);
-  }
-}
 
 function displayData(data) {
   deleteEl(contentDiv);
@@ -67,25 +63,7 @@ function displayData(data) {
   contentDiv.appendChild(div);
 }
 
-function processData(weatherData) {
-  // grab all the data i want to display on the page
-  const myData = {
-    condition: weatherData.current.condition.text,
-    feelsLike: {
-      f: Math.round(weatherData.current.feelslike_f),
-      c: Math.round(weatherData.current.feelslike_c),
-    },
-    currentTemp: {
-      f: Math.round(weatherData.current.temp_f),
-      c: Math.round(weatherData.current.temp_c),
-    },
-    wind: Math.round(weatherData.current.wind_mph),
-    humidity: weatherData.current.humidity,
-    location: weatherData.location.name.toUpperCase(),
-  };
 
-  return myData;
-}
 
 async function getData(location) {
   try {
@@ -97,16 +75,6 @@ async function getData(location) {
   } catch{
     errorhandle()
   }
-}
-
-function errorhandle() {
-  deleteEl(contentDiv)
-  const errorMsg = document.createElement('div')
-  errorMsg.classList.add('text-light')
-  errorMsg.classList.add('fs-4')
-
-  errorMsg.textContent = 'Sorry we could not find your city'
-  contentDiv.appendChild(errorMsg)
 }
 
 submit.addEventListener('click', (e) => {
